+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Added ``monitor_kwargs`` parameter
 - Handle the `env_kwargs` of `render:True` under the hood for panda-gym v1 envs in `enjoy` replay to match visualzation behavior of other envs
 - Added support for python config file
+- Added ``--tags`` argument to ``train.py`` to add tags to the wandb run
+- Added a sb3 version tag to the wandb run
 
 ### Bug fixes
 - Allow `python -m rl_zoo3.cli` to be called directly
@@ -190,6 +192,7 @@
 
 ### New Features
 - Added support for Python 3.10
+- Added 
 
 ### Bug fixes
 - Fix `--load-last-checkpoint` (@SammyRamone)